// frontend/src/shared/lib/api/model/organizationType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */

/**
 * Classification of organizations participating in compound lifecycle.
 */
export type OrganizationType = typeof OrganizationType[keyof typeof OrganizationType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrganizationType = {
  internal: 'internal',
  pharma_partner: 'pharma_partner',
  cro: 'cro',
  academic: 'academic',
  vendor: 'vendor',
  government: 'government',
} as const;